Ice Table Read Only Fix
Fixed the issue that allowed the program to make most ICE tables read only. Some of those are functional and require write access to all the users to work. Ice table field security can still be modified with the DMT section, so the power is in the users hands to actively make that decision.